src/server/requirements.txt
fastapi==0.101.0
uvicorn==0.23.2
pydantic==2.1.1
pydantic-settings==2.0.2
python-dotenv
pytest
filelock==3.13.1
fsspec==2024.3.1
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
Pillow==8.1.0
sympy==1.12
pyyaml
python-multipart
pika==1.3.2
aio_pika==9.4.1
redis==5.0.3